Make qDoc build with main. It's not quite perfect yet, but it should at least run and not seg fault.
[git-p4: depot-paths = "//depot/qt/main/": change = 106587]

+++ util/qdoc/codechunk.cpp
@@ -52,7 +52,7 @@ static const bool needSpace[7][7] = {
 
 static int category( QChar ch )
 {
-    return charCategory[ch.latin1()];
+    return charCategory[(int)(uchar)ch.latin1()];
 }
 
 CodeChunk::CodeChunk()
